requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-openai==0.1.7
python-dotenv==1.0.0
setuptools==67.7.2